requirements.txt
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
bta-lib==1.0.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
constantly==15.1.0
cryptography==3.1
dateparser==0.7.6
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
numpy==1.20.0
pandas==1.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-binance==0.7.9
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
requests==2.24.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.2.1
Twisted==20.3.0
txaio==20.4.1
tzlocal==2.1
ujson==4.0.2
urllib3==1.25.10
zope.interface==5.1.0
